pyproject.toml
[project]
name = "discord-bot-template"
version = "0.1.0"
description = "Template for a Discord bot"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "aiohttp>=3.11.18",
    "httpx>=0.28.1",
    "httpx-sse>=0.4.0",
    "pydantic>=2.11.4",
    "python-dotenv>=1.1.0",
    "discord.py>=2.5.2"
]
